Fix bug Timer
